requirements.txt
cairocffi
CairoSVG
certifi
cffi
chardet
cssselect2
defusedxml
feedparser
html5lib
idna
Pillow
pycparser
Pyphen
requests
six
tinycss2
urllib3
WeasyPrint
pdf2image
webencodings
